+++ crates/forge/src/lib.rs
@@ -1,12 +1,10 @@
 use std::collections::HashMap;
 use std::fmt::Debug;
-use std::path::PathBuf;
 
-use anyhow::{anyhow, bail, Context, Result};
+use anyhow::{anyhow, Context, Result};
 use ark_std::iterable::Iterable;
 use assert_fs::fixture::{FileTouch, PathChild, PathCopy};
 use assert_fs::TempDir;
-use cairo_felt::Felt252;
 use camino::Utf8PathBuf;
 use rayon::prelude::{IntoParallelRefIterator, ParallelIterator};
 use serde::Deserialize;
@@ -17,8 +15,6 @@ use cairo_lang_sierra::ids::ConcreteTypeId;
 use cairo_lang_sierra::program::{Function, Program};
 use cairo_lang_sierra_to_casm::metadata::MetadataComputationConfig;
 use cairo_lang_utils::ordered_hash_map::OrderedHashMap;
-use num_bigint::BigUint;
-use num_traits::Zero;
 use once_cell::sync::Lazy;
 use rand::{thread_rng, RngCore};
 use smol_str::SmolStr;
@@ -28,6 +24,7 @@ use crate::fuzzer::RandomFuzzer;
 use crate::running::run_from_test_case;
 use crate::scarb::{ForgeConfig, ForkTarget, StarknetContractArtifacts};
 pub use crate::test_crate_summary::TestCrateSummary;
+
 use test_collector::{collect_tests, LinkedLibrary, TestCase};
 
 pub mod pretty_printing;
@@ -55,7 +52,7 @@ static BUILTINS: Lazy<Vec<&str>> = Lazy::new(|| {
 });
 
 /// Configuration of the test runner
-#[derive(Deserialize, Debug, PartialEq, Default)]
+#[derive(Deserialize, Debug, PartialEq)]
 pub struct RunnerConfig {
     test_name_filter: Option<String>,
     exact_match: bool,
@@ -155,20 +152,20 @@ impl RunnerParams {
 fn collect_tests_from_package(
     package_path: &Utf8PathBuf,
     package_name: &str,
-    lib_path: &Utf8PathBuf,
-    mut linked_libraries: Vec<LinkedLibrary>,
+    package_source_dir_path: &Utf8PathBuf,
+    linked_libraries: &[LinkedLibrary],
     corelib_path: &Utf8PathBuf,
     runner_config: &RunnerConfig,
 ) -> Result<Vec<TestsFromCrate>> {
-    let tests_folder_path = package_path.join("tests");
-    let maybe_tests_tmp_dir = if tests_folder_path.try_exists()? {
+    let tests_dir_path = package_path.join("tests");
+    let maybe_tests_tmp_dir = if tests_dir_path.try_exists()? {
         Some(pack_tests_into_one_file(package_path)?)
     } else {
         None
     };
 
     let mut all_test_roots = vec![TestCrate {
-        crate_root: lib_path.clone(),
+        crate_root: package_source_dir_path.clone(),
         crate_name: package_name.to_string(),
         crate_type: TestCrateType::Lib,
     }];
@@ -176,24 +173,18 @@ fn collect_tests_from_package(
     if let Some(tests_tmp_dir) = &maybe_tests_tmp_dir {
         let tests_tmp_dir_path = Utf8PathBuf::from_path_buf(tests_tmp_dir.to_path_buf())
             .map_err(|_| anyhow!("Failed to convert tests temporary directory to Utf8PathBuf"))?;
-        let tests_lib_path = tests_tmp_dir_path.join("lib.cairo");
 
         all_test_roots.push(TestCrate {
-            crate_root: tests_lib_path,
+            crate_root: tests_tmp_dir_path.clone(),
             crate_name: "tests".to_string(),
             crate_type: TestCrateType::Tests,
-        });
-
-        linked_libraries.push(LinkedLibrary {
-            name: "tests".to_string(),
-            path: PathBuf::from(tests_tmp_dir_path),
         });
     }
 
     let tests_from_files = all_test_roots
         .par_iter()
         .map(|test_crate| {
-            collect_tests_from_tree(test_crate, &linked_libraries, corelib_path, runner_config)
+            collect_tests_from_tree(test_crate, linked_libraries, corelib_path, runner_config)
         })
         .collect();
 
@@ -244,7 +235,7 @@ fn pack_tests_into_one_file(package_path: &Utf8PathBuf) -> Result<TempDir> {
 
 fn collect_tests_from_tree(
     test_crate: &TestCrate,
-    linked_libraries: &Vec<LinkedLibrary>,
+    linked_libraries: &[LinkedLibrary],
     corelib_path: &Utf8PathBuf,
     runner_config: &RunnerConfig,
 ) -> Result<TestsFromCrate> {
@@ -300,15 +291,15 @@ pub fn run(
     package_root: &Utf8PathBuf,
     package_path: &Utf8PathBuf,
     package_name: &str,
-    lib_path: &Utf8PathBuf,
-    linked_libraries: Vec<LinkedLibrary>,
+    package_source_dir_path: &Utf8PathBuf,
+    linked_libraries: &[LinkedLibrary],
     runner_config: &RunnerConfig,
     runner_params: &RunnerParams,
 ) -> Result<Vec<TestCrateSummary>> {
     let tests = collect_tests_from_package(
         package_path,
         package_name,
-        lib_path,
+        package_source_dir_path,
         linked_libraries,
         &runner_params.corelib_path,
         runner_config,
@@ -453,25 +444,23 @@ fn run_with_fuzzing(
     case: &TestCase,
     args: &Vec<&ConcreteTypeId>,
 ) -> Result<(TestCaseSummary, u32)> {
-    if contains_non_felt252_args(args) {
-        bail!(
-            "Fuzzer only supports felt252 arguments, and test {} defines arguments that are not felt252 type",
-            case.name.as_str()
-        );
-    }
+    let args = args
+        .iter()
+        .map(|arg| {
+            arg.debug_name
+                .as_ref()
+                .ok_or_else(|| anyhow!("Type {arg:?} does not have a debug name"))
+                .map(smol_str::SmolStr::as_str)
+        })
+        .collect::<Result<Vec<_>>>()?;
 
-    let mut fuzzer = RandomFuzzer::new(
-        runner_config.fuzzer_seed,
-        runner_config.fuzzer_runs,
-        args.len(),
-        &BigUint::zero(),
-        &Felt252::prime(),
-    );
+    let mut fuzzer =
+        RandomFuzzer::create(runner_config.fuzzer_seed, runner_config.fuzzer_runs, &args)?;
 
     let mut results = vec![];
 
     for _ in 1..=runner_config.fuzzer_runs {
-        let args = fuzzer.next_felt252_args();
+        let args = fuzzer.next_args();
 
         let result = run_from_test_case(
             package_root,
@@ -497,15 +486,6 @@ fn run_with_fuzzing(
         .clone();
     let runs = u32::try_from(results.len())?;
     Ok((result, runs))
-}
-
-fn contains_non_felt252_args(args: &Vec<&ConcreteTypeId>) -> bool {
-    args.iter().any(|pt| {
-        if let Some(name) = &pt.debug_name {
-            return name != &SmolStr::from("felt252");
-        }
-        false
-    })
 }
 
 fn function_args<'a>(function: &'a Function, builtins: &[&str]) -> Vec<&'a ConcreteTypeId> {
@@ -895,29 +875,5 @@ mod tests {
                 },
             ]
         );
-    }
-
-    #[test]
-    fn args_with_only_felt252() {
-        let typ = ConcreteTypeId {
-            id: 0,
-            debug_name: Some(SmolStr::from("felt252")),
-        };
-        let args = vec![&typ, &typ];
-        assert!(!contains_non_felt252_args(&args));
-    }
-
-    #[test]
-    fn args_with_not_felt252() {
-        let typ = ConcreteTypeId {
-            id: 0,
-            debug_name: Some(SmolStr::from("felt252")),
-        };
-        let typ2 = ConcreteTypeId {
-            id: 0,
-            debug_name: Some(SmolStr::from("Uint256")),
-        };
-        let args = vec![&typ, &typ, &typ2];
-        assert!(contains_non_felt252_args(&args));
     }
 }