+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Forge
+
+#### Added
+
+- Support for `u8`, `u16`, `u32`, `u64`, `u128`, `u256` types to fuzzer
+
+#### Changed
+
+- Spying events interface is updated to enable the use of events defined inside contracts in assertions
 
 ## [0.7.1] - 2023-09-27